k parameter in RSO Hexagonal net is now read as double

+++ src/ProgramSettings.h
@@ -159,7 +159,7 @@ public:
 			double depth;
 			/*distorsion wave parameter*/
 			double alpha, frac;
-			int k;
+			double k;
 			/*how much the center of the net is moved*/
 			double center_shift;
 		};
